Feat: new `useControlled` hook
