+++ .changeset/sharp-sloths-thank.md
@@ -1,0 +1,5 @@
+---
+"react-cool-form": patch
+---
+
+Refactor(useFormState): use error instead warn for form id related exceptions
