+++ src/utils/invariant.test.ts
@@ -1,0 +1,9 @@
+import invariant from "./invariant";
+
+describe("invariant", () => {
+  it("should work correctly", () => {
+    const message = "Oops!";
+    expect(() => invariant(true, message)).toThrow(message);
+    expect(() => invariant(false, message)).not.toThrow(message);
+  });
+});
