+++ src/utils/invariant.ts
@@ -1,0 +1,3 @@
+export default (condition: boolean, message: string): void => {
+  if (__DEV__ && condition) throw new Error(message);
+};
